Cleanup of cloud provision
